requirements.txt
Django==4.2
mysqlclient==2.1.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-decouple==3.5
django-environ==0.8.1